refactor: move cli subcommands into separate modules
